Make camera_image interval configurable
Adds support for "streaming" video in the frontend by displaying still images at a rate of 1frame/stream_interval seconds (configurable in configuration.yaml). The camera_image method is only called by the frontend when the frontend is open and normally only every 10s. When the stream is opened, the camera_image method will ensure that a new image is only fetched every stream_interval seconds. Default is 5 seconds to be gentle to Kuna's servers.